requirements.txt
Flask==2.2.5
Flask-RESTful
Flask-PyMongo
Flask-JWT-Extended
flask_mongoengine
Flasgger
pydantic
pydantic-settings
python-dotenv
pytest
pytest-flask
werkzeug
mongoengine
requests
